protect posts

+++ app/Http/Controllers/Backend/CommunityController.php
@@ -52,6 +52,7 @@ class CommunityController extends Controller
      */
     public function edit(Community $community)
     {
+        $this->authorize('update', $community);
         return Inertia::render('Communities/Edit', compact('community'));
 
     }
@@ -61,7 +62,7 @@ class CommunityController extends Controller
      */
     public function update(CommunityStoreRequest $request, Community $community)
     {
-        // $this->authorize('update', $community);
+        $this->authorize('update', $community);
         $community->update($request->validated());
 
         return to_route('communities.index')->with('message', 'Community updated successfully.');
@@ -72,7 +73,7 @@ class CommunityController extends Controller
      */
     public function destroy(Community $community)
     {
-        // $this->authorize('delete', $community);
+        $this->authorize('delete', $community);
         $community->delete();
         return back()->with('message', 'Community deleted successfully.');
     }
